Major project refactoring
A lot of new methods were created, several new classes.
Main change - separate the user interface, from the game engine and from
the main game class.

+++ Hangman-7/TopPlayer.cs
@@ -1,5 +1,5 @@
 ﻿internal class TopPlayer
 {
     public string PlayerName { get; set; }
-    public double PlayerScore { get; set; }
+    public int PlayerScore { get; set; }
 }
